Actualizacion de los ultimos 3 challenges mas detalles

+++ Dia 12/Dia12_ChallengeAlgoritmo/src/main/java/com/bootcamp/Dia12_ChallengeAlgoritmo/Controller/UserController.java
@@ -2,13 +2,12 @@ package com.bootcamp.Dia12_ChallengeAlgoritmo.Controller;
 
 import com.bootcamp.Dia12_ChallengeAlgoritmo.Model.UserModel;
 import com.bootcamp.Dia12_ChallengeAlgoritmo.Service.UserService;
-import org.apache.catalina.User;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import javax.swing.text.html.Option;
-import java.util.List;
+import java.util.ArrayList;
 import java.util.Optional;
 
 @RestController
@@ -18,14 +17,28 @@ public class UserController {
     UserService userService;
 
         @GetMapping
-        public ResponseEntity<List<UserModel>> getAllUsers() {
+        public ResponseEntity<ArrayList<UserModel>> getAllUsers() {
             return userService.getAllUser();
         }
 
 
         @GetMapping("/{dni}")
-        public ResponseEntity<Optional<UserModel>> getUserByDni(@PathVariable String dni) {
+        public ResponseEntity<Optional<UserModel>> getUseByDni(@PathVariable("dni") String dni) {
+            if(dni == null){
+                return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
+            }
+            if(dni.length()<=0 ^ dni.length() >8){
+                return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
+            }
             return userService.getUserByDni(dni);
+        }
+
+        @GetMapping("/user/{name}")
+        public ResponseEntity<ArrayList<UserModel>> getUserByname(@PathVariable("name") String name) {
+            if(name == null){
+                return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
+            }
+            return userService.getUserByName(name);
         }
 
         @PostMapping()
@@ -39,7 +52,13 @@ public class UserController {
         }
 
         @DeleteMapping("/{dni}")
-        public ResponseEntity deleteUser(@PathVariable String dni){
+        public ResponseEntity deleteUser(@PathVariable("dni") String dni){
+            if(dni == null){
+                return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
+            }
+            if(dni.length()<=0 ^ dni.length() >8){
+                return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
+            }
             return userService.deleteUserByDni(dni);
         }
